Desktop/Inventory/requirements.txt
flask
flask_login
jwt
uuid
flask_wtf
flask_PyMongo
pandas
zinc
